+++ index.scss
@@ -1,6 +1,6 @@
 body {
-  padding-left: 20px;
-  padding-right: 20px;
+  // padding-left: 20px;
+  // padding-right: 20px;
 }
 
 a {
