+++ router.tsx
@@ -1,5 +1,5 @@
 import { Main } from './src/Page/Main/Main';
-import { SignIn } from './src/Page/SgnIn/SignIn';
+import { SignIn } from './src/Page/SignIn/SignIn';
 import { SignUp } from './src/Page/SignUp/SignUp';
 import { Welcome } from './src/Page/Welcome/Welcome';
 import { SIGNIN_ROUTE, SIGNUP_ROUTE, MAIN_ROUTE, WELCOME_ROUTE } from './src/utils/consts';
@@ -22,5 +22,9 @@ export const privateRoutes = [
   {
     path: MAIN_ROUTE,
     Component: <Main />,
+  },
+  {
+    path: WELCOME_ROUTE,
+    Component: <Welcome />,
   },
 ];
