+++ src/components/Form/Form.module.scss
@@ -1,0 +1,7 @@
+.container__auth {
+  margin: 100px auto;
+  width: 37%;
+  padding: 75px;
+  border:1px solid white;
+  border-radius: 5px;
+}
