+++ src/components/HeaderAuth/HeaderAuth.module.scss
