+++ src/utils/consts.tsx
@@ -1,4 +1,4 @@
 export const SIGNIN_ROUTE = '/sign-in';
-export const SIGNUP_ROUTE = '/sign-upp';
+export const SIGNUP_ROUTE = '/sign-up';
 export const MAIN_ROUTE = '/main';
 export const WELCOME_ROUTE = '/welcome';
